production environment, i change some major stuff

+++ app/Http/Controllers/Admin/PostTagController.php
@@ -11,12 +11,14 @@ class PostTagController extends Controller
     // CREATE
     public function store($postId, $tagIds){
         // create new tags AND delete all tags, that were not created
-        foreach ($tagIds as $tagId) {
-            $this->storeTag($postId, $tagId);
+        if(count($tagIds) > 0){
+            foreach ($tagIds as $tagId) {
+                $this->storeTag($postId, $tagId);
+            }
+    
+            // delete
+            $this->destroy($tagIds, $postId);
         }
-
-        // delete
-        $this->destroy($tagIds, $postId);
     }
 
     private function storeTag($postId, $tagId){
